pyproject.toml
[tool.poetry]
name = "recordmusic"
version = "0.1.0"
description = ""
authors = ["ajib6ept <asalbert@narod.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
yandex-music = "^2.0.1"
requests = "^2.28.2"
lxml = "^4.9.2"
cachelib = "^0.10.1"
python-dotenv = "^0.21.1"

[tool.poetry.dev-dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
